restructures command handling to enable encapsulation and add docs
