+++ gui/src/actions/command.rs
@@ -1,0 +1,11 @@
+use uuid::Uuid;
+
+#[derive(Debug, Clone)]
+pub(crate) enum Command {
+    SetResourceGain(usize),
+    PayResources(usize),
+    SetMultiplier(isize),
+    CloseCard(Uuid),
+    DeactivateCard(Uuid),
+    ActivateCard(Uuid),
+}
