+++ gui/src/actions/mod.rs
@@ -1,0 +1,2 @@
+pub(crate) mod command;
+pub(super) mod command_handler;
