+++ gui/src/app.rs
@@ -13,7 +13,7 @@ use game_lib::world::resources::Resources;
 use super::{Input, Message, SecCardGameApp};
 use crate::card_view_model::CardContent;
 use crate::card_window::display_card;
-use crate::command_handler::CommandHandler;
+use crate::actions::command_handler::CommandHandler;
 
 impl SecCardGameApp {
     fn init(deck: Deck) -> Self {
