+++ gui/src/lib.rs
@@ -1,7 +1,7 @@
-use crate::command_handler::Command;
 use game_lib::world::game::Game;
 use std::cell::RefCell;
 use std::rc::Rc;
+use crate::actions::command::Command;
 
 mod app;
 mod card_view_model;
@@ -9,7 +9,7 @@ mod card_window;
 pub mod start;
 mod control_panel;
 mod controls;
-mod command_handler;
+mod actions;
 
 pub(crate) type CommandToExecute = Rc<RefCell<Option<Command>>>;
